requirements.txt
PyQt5
qtwidgets
clipboard
PyCryptodome
qtpy
